refactor: extract to a separate file some helper functions
